pyproject.toml
[project]
name = "meetmind"
version = "0.1.0"
description = "Sistema de procesamiento de reuniones con IA"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "sqlmodel>=0.0.22",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "gradio>=4.0.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "langchain-openai",
    "openai>=1.40.0",
    "openai-whisper",
    "imageio-ffmpeg>=0.5.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
